+++ Assets/Scripts/Enemy/SOEnemyData.cs
@@ -5,15 +5,15 @@ using UnityEngine.UIElements;
     public class SO_EnemyData : ScriptableObject
     { 
        [SerializeField] private string _enemyName; 
-       [SerializeField] private float _enemyHealth; 
-       [SerializeField] private float _enemyDamage; 
+       [SerializeField] private int _enemyHealth; 
+       [SerializeField] private int _enemyDamage; 
        [SerializeField] private bool _enemyIsAlive;
        [SerializeField] private Color _enemyColor;
        [SerializeField] private Texture2D _texture;
    
         public string GetEnemyName => _enemyName;
-        public float GetEnemyHealth => _enemyHealth;
-        public float GetEnemyDamage => _enemyDamage;
+        public int GetEnemyHealth => _enemyHealth;
+        public int GetEnemyDamage => _enemyDamage;
         public Color GetEnemyColor => _enemyColor;
         public bool GetEnemyIsAlive => _enemyIsAlive;
